\name{PattFreq}
\alias{PattFreq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Frequency of row patterns}
\description{Compute the frequencies of row patterns in a matrix}
\usage{patf <- PattFreq(x)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Matrix of data whose row frequencies are computed}
}
\details{}
\value{patf: A list with
  \item{patrones}{A vector with the patterns in rows}
  \item{frx}{A vector with the assocaited frequencies in unit percentage points.}
}
\references{}
\author{Andres Romeu
    Fundamentos del Analisis Economico
    Universidad de Murcia 30100 SPAIN}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
