\name{TEntropy}
\alias{TEntropy}
\title{Compute total entropy of already symbolized data}
\description{Compute total entropy of already symbolized data from function "symbolize"}
\usage{entr <- TEntropy(mh)}
%- maybe also 'usage' for other objects documented here.
\arguments{Input is a list obtained from symbolized data}
\details{}
\value{entr = a structure with entropy of x, y, z, xy, xz and xyz
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Andres Romeu
    Fundamentos del Analisis Economico
    Universidad de Murcia 30100 SPAIN
}
\note{Translated from original Matlab code by my colleague Ruiz-Marin}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{symbolize}
\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
